// src/api/generated/models/materialRequestPriority.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */

export type MaterialRequestPriority = typeof MaterialRequestPriority[keyof typeof MaterialRequestPriority];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const MaterialRequestPriority = {
  LOW: 'LOW',
  NORMAL: 'NORMAL',
  HIGH: 'HIGH',
  URGENT: 'URGENT',
} as const;